+++ src/main/java/br/com/fiap/consumidorone/components/Consumer.java
@@ -1,20 +1,5 @@
 package br.com.fiap.consumidorone.components;
 
-import br.com.fiap.consumidorone.domain.BolsaFamiliaModel;
-import br.com.fiap.consumidorone.services.ConsumerService;
-import java.io.IOException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.stereotype.Component;
+public interface Consumer {
 
-@Component
-public class Consumer {
-
-    @Autowired
-    private ConsumerService service;
-
-    @KafkaListener(topics = "${consumerone.topic}", groupId = "${consumerone.group-id}")
-    public void consume(final BolsaFamiliaModel consumerRecord) {
-            service.consume(consumerRecord);
-    }
 }
